Allow setting the stop timeout for stopping a FNatsServer
When stopping a FNatsServer, the stop method awaits the server's
configured executor service for up to a defined number of seconds.
This is to give the server time to finish processing current tasks.

This time use to be hard coded to 30 seconds. Different servers may
have different needs, however. Allowing callers to both configure
the stop timeout for the server or to specify a timeout while calling
stop will enable additional use cases where a timeout other than the
default is desired.