Wait for unsubscribes to finish before shutting down executor service.
Previously, if a stop signal came in, the executor service was shutdown
prior to unsubscribing to the nats messages for the server. This would
lead to nats messages still being delivered to the server despite that
the messages will never be executed / acted upon because the executor
service was shutdown.

This changes the stop / shutdown process to first wait for the
unsubscriptions to finish prior to shutting down the executor service.
Thus, no new messages should be received from the nats server while
this frugal server is shutting down.